NGC-2391 Remove unnecessary `PushMessage` response body from `POST /messages/:id/response` endpoint
I wondered whether it was included so that the client app could see the update that it had made but acknowledging or answering a message doesn't change the `PushMessage`, the responses are stored separately.

Return an empty JSON object instead of no body as a defensive measure in case API Platform or the client requires the body to be valid JSON.

Change "201" to "200" in test name on the assumption that it was a typo.