Remove unused, non-working code
getCurrentMessages and respondToMessage disagreed about the meaning of the Acknowledge and Answer states - getCurrentMessages interpreted Acknowledge to mean "NOT YET acknowledged" and respondToMessage interpreted it to mean "HAS BEEN acknowledged".

Rather than fix this incompatibility we're just removing getCurrentMessages because it is not currently used.

Original JIRA for the code that was removed: NGC-1878.

Added comments to Acknowledge and Answer states documenting the way they are used by the remaining code (these descriptions copied from the "Release 2 overview" Google doc).